Add fractional difficulty support to solo-ckpool
- Modified ckpool.h: Changed difficulty fields from int64_t to double
- Updated ckpool.c: Use json_get_double() for configuration parsing
- Enhanced stratifier.c: Removed integer-only validation constraints
- Maintains backward compatibility with integer configurations
- Enables support for fractional difficulty values < 1.0

+++ src/ckpool.h
@@ -231,10 +231,10 @@ struct ckpool_instance {
 	int nonce2length; // Extranonce2 length
 
 	/* Difficulty settings */
-	int64_t mindiff; // Default 1
-	int64_t startdiff; // Default 42
-	int64_t highdiff; // Default 1000000
-	int64_t maxdiff; // No default
+	double mindiff; // Default 1.0 (supports fractional values)
+	double startdiff; // Default 42.0 (supports fractional values)
+	double highdiff; // Default 1000000.0 (supports fractional values)
+	double maxdiff; // No default (supports fractional values)
 
 	/* Coinbase data */
 	char *btcaddress; // Address to mine to
